Add views array to XRViewerPose.
This also requires implementing part of XRRigidTransform and
adding a transform to XRView.

As of this commit, I have the xr-presentation.html sample from
https://storage.googleapis.com/chromium-webxr-test/r672724/xr-presentation.html
working with this polyfill when hosting the samples locally. The only
change I had to make to the html file was how the polyfill gets injected.

I tested in the following browsers on Windows using an Oculus headeset:
- Chrome 75 with WebVR flag enabled, but WebXR flag disabled
- Firefox 67.0.4