Expose input poses and gamepad from WebVR
Made sure that the data exposed to gamepads via the WebVR API is
properly forwarded to the WebXR API surface. Includes both the poses
and repeating the original gamepad object on the XRInputSource.gamepad
attribute. No remapping of the gamepad API data is being done at this
point. Removed the XRInputPose, because it's no longer part of the API.