Allow mapping of WebVR gamepad values
Provide a mapping table that allows WebVR-style gamepads to be remapped
so that they can conform to the WebXR conventions and the `xr-standard`
mapping. Also allows for static transforms to be applied to either the
grip or the targetRay poses so that the single value that WebVR supplied
can be used for both spaces accurately.